+++ scripts/mame-pin.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+readonly MAME_TAG="mame0289"
+readonly MAME_COMMIT="f34f02505e32c1993c6a782b6814232cbfc74e36"
+readonly MAME_ARCHIVE_SHA256="17d50a6effe503e5cd23818daf42ee2a60f471d1cda41c13e0e7cc4ae78c5e11"
+readonly MAME_ARCHIVE_URL="https://github.com/mamedev/mame/archive/${MAME_COMMIT}.tar.gz"